% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_file.R
\name{OpenHelpFile}
\alias{OpenHelpFile}
\title{Open MSPepSearch help file}
\usage{
OpenHelpFile()
}
\value{
Return \code{NULL}.
}
\description{
Open the \emph{MSPepSearch64.exe.hlp.txt} file included with the
  MSPepSearch (NIST) tool in the default text editor.
}
